feat:방 기능 구현 [BACK-124]

Closes BACK-124

See merge request s09-webmobile1-sub2/S09P12A307!48

+++ backend/src/main/java/signiel/heartsigniel/model/party/PartyRequest.java
@@ -21,5 +21,5 @@ public class PartyRequest {
 
     private String partyType;
 
-    private int avgRating;
+    private Long avgRating;
 }
